refactor: move Question schemas to his own file and unify canonical Question schema adding dataset_id (#4487)
# Description

This PR includes the following changes:
* Move `Question` related schemas for API v1 to his own file at
`schemas/v1/questions.py`.
* Use only one `Question` canonical schema removing duplications.
* `Question` schema will include always `dataset_id` as attribute.
* Small renaming of schemas to avoid conflicts in the same file.

Refs #4407 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- [ ] Running and modifying unit tests.

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)
